adding missing files, extraced classes into files
