+++ .travis.yml
@@ -1,0 +1,34 @@
+language: node_js
+node_js: "10.15.0"
+os: linux
+services:
+  - xvfb
+before_install:
+  - 'export PATH=./node_modules/.bin:$PATH'
+  - 'npm install -g typescript'
+  - 'npm install -g vsce'
+install:
+  - 'npm install --ignore-scripts'
+  - 'npm install'
+  # https://github.com/travis-ci/travis-ci/issues/8813
+  - 'rm -f ./node_modules/.bin/which'
+  - 'vsce package'
+script:
+  - 'npm test --silent'
+  - 'npm run ui-test'
+after_success:
+  - if [[ $TRAVIS_PULL_REQUEST == "false" && $TRAVIS_BRANCH == "master" ]]; then
+    sonar-scanner;
+    fi
+cache:
+  directories:
+    - "node_modules"
+env: DISPLAY=:99
+branches:
+  except:
+    - /^dependabot.*$/
+addons:
+  sonarcloud:
+    organization: "jboss-fuse"
+    branches:
+      - master
