+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+	"editor.insertSpaces": false,
+	"editor.detectIndentation": false,
+	"prettier.tabWidth": 4,
+	"prettier.useTabs": true
+}
