+++ .vscodeignore
@@ -1,0 +1,11 @@
+**/*.ts
+*.vsix
+.gitignore
+.travis.yml
+.vscode/**
+gulpfile.js
+sonar-project.properties
+tsconfig.json
+tslint.json
+Contributing.md
+test-resources
