+++ package.json
@@ -144,15 +144,8 @@
 		"tslint": "gulp tslint"
 	},
 	"dependencies": {
-		"child_process": "^1.0.2",
-		"detect-port": "^1.3.0",
-		"expand-home-dir": "^0.0.3",
-		"fetch": "^1.1.0",
-		"find-java-home": "^1.1.0",
 		"fs": "^0.0.1-security",
 		"path": "^0.12.7",
-		"path-exists": "^4.0.0",
-		"util": "^0.12.1",
 		"yaml": "^1.7.2"
 	},
 	"devDependencies": {
